+++ octodns/provider/base.py
@@ -95,7 +95,50 @@ class BaseProvider(BaseSource):
                 record.values = [record.value]
                 desired.add_record(record, replace=True)
 
+        record = desired.root_ns
+        if self.SUPPORTS_ROOT_NS:
+            if not record:
+                self.log.warning('root NS record supported, but no record '
+                                 'is configured for %s', desired.name)
+        else:
+            if record:
+                # we can't manage root NS records, get rid of it
+                msg = \
+                    f'root NS record not supported for {record.fqdn}'
+                fallback = 'ignoring it'
+                self.supports_warn_or_except(msg, fallback)
+                desired.remove_record(record)
+
         return desired
+
+    def _process_existing_zone(self, existing, desired):
+        '''
+        An opportunity for providers to modify the existing zone records before
+        planning. `existing` is a "shallow" copy, see `Zone.copy` for more
+        information
+
+        - `desired` must not be modified in anyway, it is only for reference
+        - Must call `super` at an appropriate point for their work, generally
+          that means as the final step of the method, returning the result of
+          the `super` call.
+        - May modify `existing` directly.
+        - Must not modify records directly, `record.copy` should be called,
+          the results of which can be modified, and then `Zone.add_record` may
+          be used with `replace=True`.
+        - May call `Zone.remove_record` to remove records from `existing`.
+        - Must call supports_warn_or_except with information about any changes
+          that are made to have them logged or throw errors depending on the
+          provider configuration.
+        '''
+
+        existing_root_ns = existing.root_ns
+        if existing_root_ns and (not self.SUPPORTS_ROOT_NS or
+                                 not desired.root_ns):
+            self.log.info('root NS record in existing, but not supported or '
+                          'not configured; ignoring it')
+            existing.remove_record(existing_root_ns)
+
+        return existing
 
     def _include_change(self, change):
         '''
@@ -120,13 +163,6 @@ class BaseProvider(BaseSource):
     def plan(self, desired, processors=[]):
         self.log.info('plan: desired=%s', desired.name)
 
-        # Make a (shallow) copy of the desired state so that everything from
-        # now on (in this target) can modify it as they see fit without
-        # worrying about impacting other targets.
-        desired = desired.copy()
-
-        desired = self._process_desired_zone(desired)
-
         existing = Zone(desired.name, desired.sub_zones)
         exists = self.populate(existing, target=True, lenient=True)
         if exists is None:
@@ -134,6 +170,15 @@ class BaseProvider(BaseSource):
             # information
             self.log.warning('Provider %s used in target mode did not return '
                              'exists', self.id)
+
+        # Make a (shallow) copy of the desired state so that everything from
+        # now on (in this target) can modify it as they see fit without
+        # worrying about impacting other targets.
+        desired = desired.copy()
+
+        desired = self._process_desired_zone(desired)
+
+        existing = self._process_existing_zone(existing, desired)
 
         for processor in processors:
             existing = processor.process_target_zone(existing, target=self)