+++ tests/test_octodns_manager.py
@@ -216,14 +216,16 @@ class TestManager(TestCase):
             with open(join(tmpdir.dirname, 'unit.tests.yaml'), 'w') as fh:
                 fh.write('---\n{}')
 
+            # compare doesn't use _process_desired_zone and thus doesn't filter
+            # out root NS records, that seems fine/desirable
             changes = manager.compare(['in'], ['dump'], 'unit.tests.')
-            self.assertEqual(20, len(changes))
+            self.assertEqual(21, len(changes))
 
             # Compound sources with varying support
             changes = manager.compare(['in', 'nosshfp'],
                                       ['dump'],
                                       'unit.tests.')
-            self.assertEqual(19, len(changes))
+            self.assertEqual(20, len(changes))
 
             with self.assertRaises(ManagerException) as ctx:
                 manager.compare(['nope'], ['dump'], 'unit.tests.')