+++ tests/test_octodns_source_tinydns.py
@@ -29,10 +29,15 @@ class TestTinyDnsFileSource(TestCase):
                 'ttl': 30,
                 'values': ['10.2.3.4', '10.2.3.5'],
             }),
+            ('', {
+                'type': 'NS',
+                'ttl': 3600,
+                'values': ['ns1.ns.com.', 'ns2.ns.com.'],
+            }),
             ('sub', {
                 'type': 'NS',
                 'ttl': 30,
-                'values': ['ns1.ns.com.', 'ns2.ns.com.'],
+                'values': ['ns3.ns.com.', 'ns4.ns.com.'],
             }),
             ('www', {
                 'type': 'A',