+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 ## v0.9.16 - 2022-??-?? - ???
 
+#### Noteworthy changes
+
+* Foundational support for root NS record management.
+   * YamlProvider has it enabled and in general everyone should add root NS
+     records that match what is in their provider(s) as of this release if they
+     aren't already there.
+   * Note that if you created your config files with `octodns-dump`, the records
+     are likely already there and match what was configured at the time of the
+     dump.
+   * Other providers will add root NS support over time once they have had the
+     chance to investigate the functionality and implement management if
+     possible with whatever accomidations are required.
+   * Watch your providers README.md and CHANGELOG.md for support and more
+     information.
+
 #### Stuff
 
 * _AggregateTarget has more complete handling of SUPPORTS* functionality,